+++ exercicios/losa-nascimento/para-casa/projeto-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "projeto-um",
+  "name": "projeto-react",
   "private": true,
   "version": "0.0.0",
   "type": "module",
