+++ exercicios/losa-nascimento/para-casa/projeto-react/src/components/Footer.jsx
@@ -1,0 +1,9 @@
+function Footer(props) {
+    return (
+      <footer>
+        <p>Feito com ♡ por {props.name}. E todos os direitos são reservados.</p>
+      </footer>
+    )
+  }
+  
+  export default Footer
