+++ exercicios/losa-nascimento/para-casa/projeto-react/src/components/Header.jsx
@@ -1,0 +1,9 @@
+function Header() {
+    return (
+      <header>
+        <h1>Meus conteúdos favoritos</h1>
+      </header>
+    )
+  }
+  
+  export default Header
