+++ exercicios/losa-nascimento/para-casa/projeto-react/src/components/Image.jsx
@@ -1,0 +1,7 @@
+function Image(props) {
+    return (
+      <img src={props.img} alt={props.description} />
+    )
+  }
+  
+  export default Image
