+++ exercicios/losa-nascimento/para-casa/projeto-react/src/components/Subtitle.jsx
@@ -1,0 +1,7 @@
+function Subtitle(props) {
+    return (
+      <h4>{props.text}</h4>
+    )
+  }
+  
+  export default Subtitle
