+++ exercicios/losa-nascimento/para-casa/projeto-react/src/components/Text.jsx
@@ -1,0 +1,7 @@
+function Text(props) {
+    return (
+      <p>{props.text}</p>
+    )
+  }
+  
+  export default Text
