+++ exercicios/losa-nascimento/para-casa/projeto-react/src/components/Title.jsx
@@ -1,0 +1,7 @@
+function Title(props) {
+    return (
+      <h2>{props.text}</h2>
+    )
+  }
+  
+  export default Title
