--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Bio.jsx
@@ -1,17 +1,0 @@
-import Title from "./Title"
-import Text from "./Text"
-import Image from "./Image"
-
-import lili from '../assets/lilipalhacita.jpeg'
-
-function Bio() {
-  return (
-    <div className="card">
-      <Title text="Lilit Bandeira"/>
-      <Text text="Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."/>
-      <Image img={lili} description="lilit"/>
-    </div>
-  )
-}
-
-export default Bio
