--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Footer.jsx
@@ -1,9 +1,0 @@
-function Footer(props) {
-  return (
-    <footer>
-      <p>Feito com ♡ por {props.name}. E todos os direitos são reservados.</p>
-    </footer>
-  )
-}
-
-export default Footer
