--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Header.jsx
@@ -1,9 +1,0 @@
-function Header() {
-  return (
-    <header>
-      <h1>Meus conteúdos favoritos</h1>
-    </header>
-  )
-}
-
-export default Header
