--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Image.jsx
@@ -1,7 +1,0 @@
-function Image(props) {
-  return (
-    <img src={props.img} alt={props.description} />
-  )
-}
-
-export default Image
