--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Subtitle.jsx
@@ -1,7 +1,0 @@
-function Subtitle(props) {
-  return (
-    <h4>{props.text}</h4>
-  )
-}
-
-export default Subtitle
