--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Text.jsx
@@ -1,7 +1,0 @@
-function Text(props) {
-  return (
-    <p>{props.text}</p>
-  )
-}
-
-export default Text
