--- exercicios/nome-aluna/para-casa/projeto-um/src/components/Title.jsx
@@ -1,7 +1,0 @@
-function Title(props) {
-  return (
-    <h2>{props.text}</h2>
-  )
-}
-
-export default Title
